Split capcom.go based on functionality focus
Split based on the focus of the functions, be it permission handling,
SecurityGroup manipulation or general helpers.
